+++ tests/04-select.ts
@@ -1,7 +1,7 @@
 import { NanoSQLInstance } from "../src/index";
 import { expect, assert } from "chai";
 import "mocha";
-import { usersDB, ExampleUsers, ExampleDataModel } from "./data";
+import { usersDB, ExampleUsers, ExampleUsers2, ExampleDataModel } from "./data";
 
 describe("Select", () => {
     it("Select single column.", (done: MochaDone) => {
@@ -56,6 +56,41 @@ describe("Select", () => {
                             {"posts.length": 1, "posts[1]": undefined},
                             {"posts.length": 3, "posts[1]": 2}
                         ], "Select inner value failed!");
+                        done();
+                    } catch (e) {
+                        done(e);
+                    }
+                });
+            });
+        });
+    });
+    it("Secondary index range select.", (done: MochaDone) => {
+        usersDB(ExampleDataModel, (nSQL) => {
+            nSQL.loadJS("users", ExampleUsers2).then(() => {
+                nSQL.table("users").query("select", ["id", "age"]).where(["age", ">", 23]).exec().then((rows) => {
+                    try {
+                        expect(rows).to.deep.equal([
+                            { id: 2, age: 24 },
+                            { id: 6, age: 29 },
+                            { id: 5, age: 30 }
+                        ], "Secondary index range failed!");
+                        done();
+                    } catch (e) {
+                        done(e);
+                    }
+                });
+            });
+        });
+    });
+    it("Secondary index range select (part 2).", (done: MochaDone) => {
+        usersDB(ExampleDataModel, (nSQL) => {
+            nSQL.loadJS("users", ExampleUsers2).then(() => {
+                nSQL.table("users").query("select", ["id", "age"]).where(["age", "<", 23]).exec().then((rows) => {
+                    try {
+                        expect(rows).to.deep.equal([
+                            { id: 1, age: 20 },
+                            { id: 3, age: 21 }
+                        ], "Secondary index range failed!");
                         done();
                     } catch (e) {
                         done(e);