+++ tests/05-where.ts
@@ -76,8 +76,8 @@ describe("Where", () => {
                 nSQL.table("users").query("select").where(["age", ">=", 21]).exec().then((rows) => {
                     try {
                         expect(rows).to.deep.equal([
-                            {id: 2, name: "Jeb", age: 24, email: "jeb@gmail.com", meta: {value: 1}, posts: [1]},
-                            {id: 3, name: "Bob", age: 21, email: "bob@gmail.com", meta: {value: 1}, posts: [1, 2, 3]}
+                            {id: 3, name: "Bob", age: 21, email: "bob@gmail.com", meta: {value: 1}, posts: [1, 2, 3]},
+                            {id: 2, name: "Jeb", age: 24, email: "jeb@gmail.com", meta: {value: 1}, posts: [1]}
                         ], "Multiple row select failed!");
                         done();
                     } catch (e) {
@@ -130,8 +130,8 @@ describe("Where", () => {
                 nSQL.table("users").query("select").where(["age", "BETWEEN", [19, 21]]).exec().then((rows) => {
                     try {
                         expect(rows).to.deep.equal([
-                            {id: 1, name: "Bill", age: 20, email: "bill@gmail.com", meta: {value: 1}, posts: [1, 3]},
-                            {id: 3, name: "Bob", age: 21, email: "bob@gmail.com", meta: {value: 1}, posts: [1, 2, 3]}
+                            {id: 3, name: "Bob", age: 21, email: "bob@gmail.com", meta: {value: 1}, posts: [1, 2, 3]},
+                            {id: 1, name: "Bill", age: 20, email: "bill@gmail.com", meta: {value: 1}, posts: [1, 3]}
                         ], "Multiple row select with IN failed!");
                         done();
                     } catch (e) {